+++ lib/deployer/ruby_deployer.rb
@@ -22,8 +22,14 @@ class Deployer::RubyDeployer < Deployer::BaseDeployer
         puts
         puts "Deploying #{processor_name}".light_blue.on_blue
         log_deployment(executor, "Deploying #{processor_name} via #{deploy_via} from #{cached_copy_dir}")
+
         # stop old one
-        executor.execute("kill -s TERM $(cat #{File.join(pids_dir, processor_name)}.pid)", abort_on_error: false, comment: "This is not sooo important")
+        if inspector?
+          cmd = inspector_command('stop', processor_name)
+          executor.execute(cmd, abort_on_error: false, comment: "Request to stop component")
+        else
+          executor.execute("kill -s TERM $(cat #{File.join(pids_dir, processor_name)}.pid)", abort_on_error: false, comment: "This is not sooo important")
+        end
 
         # unzip package
         target = deploy_dir('ruby')
@@ -46,7 +52,12 @@ class Deployer::RubyDeployer < Deployer::BaseDeployer
         executor.execute("cd #{processor_dir(processor_name)} && bundle install --without test")
 
         # start new one
-        executor.execute("cd #{processor_dir(processor_name)} && bundle exec ruby #{processor_name}.rb -d -e $EH_ENV")
+        if inspector?
+          cmd = inspector_command('start', processor_name)
+          executor.execute(cmd, abort_on_error: false, comment: "Request to stop component")
+        else
+          executor.execute("cd #{processor_dir(processor_name)} && bundle exec ruby #{processor_name}.rb -d -e $EH_ENV")
+        end
       end
     end
   end
